delete files, add navigation
