+++ PodcastApp/AuthorizationScreen/LoginScreen/LoginViewController.swift
@@ -8,12 +8,113 @@
 import UIKit
 
 class LoginViewController: UIViewController {
+    
+    let cornerRadius: CGFloat = 24
+    
+    private lazy var loginField: InputField = {
+        let emailField = InputField(inputField: UITextField(hasBorder: false, backgroundColor: .lightGray, cornerRadius: cornerRadius, placeholder: "Enter your login"), title: "Login")
+        return emailField
+    }()
 
+    private lazy var emailField: InputField = {
+        let emailField = InputField(inputField: UITextField(hasBorder: false, backgroundColor: .lightGray, cornerRadius: cornerRadius, placeholder: "Enter your email address"), title: "Email")
+        return emailField
+    }()
+
+    private lazy var emailButton: UIButton = {
+        let button = UIButton(title: "Enter",
+                              backgroundColor: .activeBlueColor,
+                              titleColor: .white,
+                              hasBorder: false,
+                              cornerRadius: cornerRadius)
+        return button
+    }()
+
+    private lazy var dividerView: Divider = {
+        let view = Divider(title: "Or continue with")
+        return view
+    }()
+
+    private lazy var googleButton: UIButton = {
+        let button = UIButton(title: "  Continue with Google",
+                              backgroundColor: .clear,
+                              titleColor: .blackTextColor,
+                              hasBorder: true,
+                              cornerRadius: cornerRadius)
+        button.setupGoogleImage()
+        return button
+    }()
+
+    private lazy var bottomText = UILabel(name: "Don't you have an account yet?",
+                            font: .jakarta16semibold(), textColor: .boldGrayTextColor)
+
+    private lazy var registerButton: UIButton = {
+        let button = UIButton(type: .system)
+        button.setTitle(" Register", for: .normal)
+        button.titleLabel?.font = .jakarta16semibold()
+        button.setTitleColor(.greenTextColor, for: .normal)
+        button.addTarget(self, action: #selector(didTapRegisterButton), for: .touchUpInside)
+        return button
+    }()
+    
+    let bottomStackView: UIStackView = {
+        let stack = UIStackView()
+        stack.axis = .horizontal
+        return stack
+    }()
+    
     override func viewDidLoad() {
         super.viewDidLoad()
-
-        
+        setupView()
+        setConstraint()
     }
-
-
+    
+    func setupView() {
+        view.backgroundColor = .white
+        
+        bottomStackView.addArrangedSubview(bottomText)
+        bottomStackView.addArrangedSubview(registerButton)
+        
+        view.addSubviews(loginField, emailField, emailButton, dividerView, googleButton, bottomStackView)
+    }
+    
+    func setConstraint(){
+        
+        NSLayoutConstraint.activate([
+            loginField.topAnchor.constraint(equalTo: view.topAnchor, constant: 100),
+            loginField.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 24),
+            loginField.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -24),
+            loginField.heightAnchor.constraint(equalToConstant: 88),
+            
+            emailField.topAnchor.constraint(equalTo: loginField.bottomAnchor, constant: 12),
+            emailField.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 24),
+            emailField.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -24),
+            emailField.heightAnchor.constraint(equalToConstant: 88),
+            
+            emailButton.topAnchor.constraint(equalTo: emailField.bottomAnchor, constant: 24),
+            emailButton.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 24),
+            emailButton.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -24),
+            emailButton.heightAnchor.constraint(equalToConstant: 56),
+            
+            dividerView.topAnchor.constraint(equalTo: emailButton.bottomAnchor, constant: 32),
+            dividerView.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 48),
+            dividerView.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -48),
+            dividerView.heightAnchor.constraint(equalToConstant: 22),
+            
+            googleButton.topAnchor.constraint(equalTo: dividerView.bottomAnchor, constant: 32),
+            googleButton.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 24),
+            googleButton.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -24),
+            googleButton.heightAnchor.constraint(equalToConstant: 56),
+            
+            bottomStackView.bottomAnchor.constraint(equalTo: view.bottomAnchor, constant: -90),
+            bottomStackView.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+        ])
+    }
+    
+    @objc
+    private func didTapRegisterButton() {
+        let createAcc = CreateAccountViewController()
+        navigationController?.pushViewController(createAcc, animated: true)
+    }
+    
 }
